+++ SolidWorksAPI/Feature/Axis_3_Milling/Axis3_SurfaceFinishMilling.cs
@@ -102,7 +102,7 @@ namespace SolidWorksAPI
             this.TotalTime = this.AtcTime + this.OtherTime + this.CuttingTime;
         }
 
-        protected override int GetCuttingSpeed()
+        protected override double GetCuttingSpeed()
         {
             switch (this._Materials)
             {
